more tests paired

+++ src/ganon-classify/CommandLineParser.cpp
@@ -17,8 +17,8 @@ std::optional< Config > CommandLineParser::parse( int argc, char** argv )
         ( "g,group-bin", "Tab-separated file[s] mapping classification groups and bin identifiers with the following fields: group_id <tab> bin_id (e.g. a.map,b.map)", cxxopts::value< std::vector< std::string > >() )
         ( "c,filter-hierarchy", "Hierarchy of the given filter/group-bin files (e.g. 1,1,x,z)", cxxopts::value< std::vector< std::string > >() )
         ( "m,min-kmers", "Minimum percentage of k-mers matching for a read to to be assigned [muttualy exclusive --max-error]. Default: 0.25", cxxopts::value< std::vector< float > >() )
-        ( "e,max-error", "Maximum number of errors/mismatches allowed [muttualy exclusive --min-kmers]", cxxopts::value< std::vector< uint32_t > >() )
-        ( "u,max-error-unique", "Maximum number of errors/mismatches allowed for unique matches after filtering. Matches not passing this criterial will be flagged with negative k-mer counts.", cxxopts::value< std::vector< int32_t > >() )
+        ( "e,max-error", "Maximum number of errors/mismatches allowed [muttualy exclusive --min-kmers]", cxxopts::value< std::vector< uint16_t > >() )
+        ( "u,max-error-unique", "Maximum number of errors/mismatches allowed for unique matches after filtering. Matches not passing this criterial will be flagged with negative k-mer counts.", cxxopts::value< std::vector< int16_t > >() )
         ( "i,paired-mode", "Paired-end mode [1: concat]. Default: 1", cxxopts::value< int16_t >() )
         ( "f,offset", "Offset for skipping k-mers while counting. Function must be enabled on compilation time with -DGANON_OFFSET=ON. Default: 1 = no offset", cxxopts::value< uint16_t >() )
         ( "o,output-file", "Output file with classification (omit for STDOUT). ", cxxopts::value< std::string >() )
@@ -68,7 +68,7 @@ std::optional< Config > CommandLineParser::parse( int argc, char** argv )
     if ( args.count( "output-unclassified-file" ) )
         config.output_unclassified_file = args["output-unclassified-file"].as< std::string >();
     if ( args.count( "max-error" ) )
-        config.max_error = args["max-error"].as< std::vector< uint32_t > >();
+        config.max_error = args["max-error"].as< std::vector< uint16_t > >();
     if ( args.count( "min-kmers" ) )
         config.min_kmers = args["min-kmers"].as< std::vector< float > >();
     if ( args.count( "offset" ) )
@@ -88,7 +88,7 @@ std::optional< Config > CommandLineParser::parse( int argc, char** argv )
     if ( args.count( "filter-hierarchy" ) )
         config.filter_hierarchy = args["filter-hierarchy"].as< std::vector< std::string > >();
     if ( args.count( "max-error-unique" ) )
-        config.max_error_unique = args["max-error-unique"].as< std::vector< int32_t > >();
+        config.max_error_unique = args["max-error-unique"].as< std::vector< int16_t > >();
     if ( args.count( "n-batches" ) )
         config.n_batches = args["n-batches"].as< uint32_t >();
     if ( args.count( "n-reads" ) )
